+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ﻿# Changelog
 
+## [Unreleased]
+
 ## [1.1.2] - 2025-06-28
 ### Changed
 - "Show Items View Distance" option is now disabled by default (search radius is more impactful for foraging)